fix(tests): use Float for moveJoystick value to avoid Float→Double precision loss
Double(Float(0.8)) ≈ 0.800000012, not 0.8 — exact equality assertions
would always fail. Keep magnitude as Float throughout MockVectrexResponder
so test comparisons are numerically exact.

+++ PVUI/Tests/PVUIBaseTests/CompanionInputEventTests.swift
@@ -102,7 +102,7 @@ private final class MockVectrexResponder {
     enum Call: Equatable {
         case push(button: Int)
         case release(button: Int)
-        case moveJoystick(direction: Int, value: Double)
+        case moveJoystick(direction: Int, value: Float)
     }
 
     var calls: [Call] = []
@@ -133,7 +133,7 @@ private final class MockVectrexResponder {
     }
 
     private func handleAxis(_ axis: CompanionAxisID, value: Float) {
-        let magnitude = Double(abs(value))
+        let magnitude = abs(value)
         switch axis {
         case .leftX:
             if value >= 0 {
